fix(*): cleanups & refactoring
