+++ .env.example
@@ -1,6 +1,4 @@
 BACKEND_URL=http://localhost:8000
-
 GUARDRAILS_URL = http://localhost:8001
 GUARDRAILS_TOKEN = 
-
 NEXT_PUBLIC_GOOGLE_CLIENT_ID=your-google-client-id.apps.googleusercontent.com
