+++ app/api/guardrails/ban_lists/route.ts
@@ -1,12 +1,11 @@
-import { guardrailsClient } from "@/app/lib/guardrailsClient";
+import { guardrailsUserClient } from "@/app/lib/guardrailsClient";
 import { NextResponse, NextRequest } from "next/server";
 
 export async function GET(request: NextRequest) {
   try {
-    const { status, data } = await guardrailsClient(
+    const { status, data } = await guardrailsUserClient(
       request,
       "/api/v1/guardrails/ban_lists",
-      { skipEnvToken: true },
     );
     return NextResponse.json(data, { status });
   } catch (e: unknown) {
@@ -20,13 +19,12 @@ export async function GET(request: NextRequest) {
 export async function POST(request: NextRequest) {
   try {
     const body = await request.json();
-    const { status, data } = await guardrailsClient(
+    const { status, data } = await guardrailsUserClient(
       request,
       "/api/v1/guardrails/ban_lists",
       {
         method: "POST",
         body: JSON.stringify(body),
-        skipEnvToken: true,
       },
     );
     return NextResponse.json(data, { status });
