+++ app/lib/utils/guardrails.ts
@@ -1,4 +1,5 @@
 import { NextRequest } from "next/server";
+import type { ValidatorConfigSchema } from "@/app/lib/types/guardrails";
 
 /**
  * Builds the backend endpoint for validator configs, forwarding
@@ -19,4 +20,33 @@ export function buildValidatorConfigEndpoint(
     ? `/api/v1/guardrails/validators/configs/${config_id}`
     : `/api/v1/guardrails/validators/configs`;
   return `${base}${qs ? `?${qs}` : ""}`;
+}
+
+/**
+ * Resolves a `$ref` pointer like "#/$defs/GuardrailOnFail" against the
+ * provided `$defs` map and returns its `enum` values (or an empty array).
+ */
+export function resolveRef(
+  ref: string,
+  defs: ValidatorConfigSchema["$defs"],
+): string[] {
+  const key = ref.replace("#/$defs/", "");
+  return defs?.[key]?.enum ?? [];
+}
+
+/**
+ * Builds the initial form values from a validator's config schema by
+ * pulling each property's `default` (skipping `type`).
+ */
+export function buildDefaultValues(
+  schema: ValidatorConfigSchema,
+): Record<string, unknown> {
+  const values: Record<string, unknown> = {};
+  for (const [key, prop] of Object.entries(schema.properties)) {
+    if (key === "type") continue;
+    if ("default" in prop) {
+      values[key] = prop.default;
+    }
+  }
+  return values;
 }
